ai/bots/requirements.txt
streamlit==1.36.0
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langchain-openai==0.0.6
langchain-google-genai>=1.0.0,<3.0.0
psycopg2-binary==2.9.3
groq==0.4.2
langchain-groq==0.0.1
python-dotenv==1.0.0
matplotlib==3.5.1
pandas==1.4.1
seaborn==0.11.2
altair==5.3.0
sshtunnel==0.4.0
tabulate==0.9.0
